[Parse] Split out IfConfig related functions into ParseIfConfig.cpp
